Still getting weird connection closed remotely errors. Three-way fix.
(1) Use the referer url properly (the enclosing page) in case Twitter is detecting that's missing and shutting us down

(2) Use normal looking browser user-agent and other request headers in case Twitter's detecting we're a program not a browser and therefore shutting us down.

(3) Use Apache HTTP Client instead of the Netty AsyncHttpClient in case it's the fault of the latter that we're getting errors, or in case the latter is masking the true cause of the error.