Fixed errors in POM
This involved forcing the mg4j version to 5.2. I'm not sure that this
is used for actually - I feel there may be quite some cruft in the
repo. Also, something, maybe Eclipse, added classes to the git-ignore
directory which would be better if it weren't in the target directory,
which should be ignored.