Fix URL tokenization by using latest URL regexps
This is a weird one: twitter has opensources the latest version of the
"text" library with corrected regexps (version 1.6.1). However the
commons-text library is still linking against version 1.4.0 of "text"
which has caused the issue.

Unfortunately we can't straightforwardly update the dependency in the
POM as it causes link errors due to missing symbols.